pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "vplan"
version = "0.4.0"
description = "OpenClaw Vacation Planner CLI — points optimization, award search, trip research"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
    "playwright-stealth>=2.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
]

[project.scripts]
vplan = "vplan_cli.cli:main"
